Move Monix tests to tests module
